+++ projects/options/src/app/app.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core'
-import { MatLegacySlideToggleChange as MatSlideToggleChange } from '@angular/material/legacy-slide-toggle'
+import { MatSlideToggleChange } from '@angular/material/slide-toggle'
 import { ChromeSharedOptionsService } from 'chrome-shared-options'
 
 function fixWrongInputDate(wrongDate: Date): Date {
