+++ src/app/app.component.ts
@@ -2,7 +2,7 @@ import { Component, OnInit } from '@angular/core'
 import { UntypedFormControl } from '@angular/forms'
 import { Observable } from 'rxjs'
 import { map, switchMap, tap } from 'rxjs/operators'
-import { MatLegacySelectionListChange as MatSelectionListChange } from '@angular/material/legacy-list'
+import { MatSelectionListChange } from '@angular/material/list'
 import { ChromeService } from './chrome.service'
 import Fuse from 'fuse.js'
 import { ChromeSharedOptionsService } from 'chrome-shared-options'
